BackEnd/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
bidict==0.23.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
DBUtils==3.1.2
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
dotenv==0.9.9
emoji==2.15.0
Faker==40.4.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kombu==5.6.2
mongoengine==0.29.1
multidict==6.7.1
openai==2.21.0
packaging==26.0
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
redis==7.2.0
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
wsproto==1.3.2
yarl==1.22.0
zapi==0.0.8
